rename files, use better read & write methods
