+++ src/hp/derp.rs
@@ -1,7 +1,7 @@
+mod client;
 mod derp;
-mod derp_client;
-mod derp_map;
 pub mod http;
+mod map;
 
-pub use derp_client::ReceivedMessage;
-pub use derp_map::{DerpMap, DerpNode, DerpRegion, UseIpv4, UseIpv6};
+pub use client::ReceivedMessage;
+pub use map::{DerpMap, DerpNode, DerpRegion, UseIpv4, UseIpv6};
