+++ src/hp/derp/client.rs
