+++ src/hp/derp/map.rs
